requirements.txt
fastapi
uvicorn
python-dotenv
httpx
beautifulsoup4
playwright
feedparser
ddgs
pytrends
groq
libsql-experimental
bcrypt
itsdangerous
amadeus
python-multipart
python-telegram-bot
huggingface_hub
pytest
pytest-asyncio
